fix: fix authorization issues

+++ src/shared/db/graphql/resolvers/label.resolvers.ts
@@ -1,52 +1,57 @@
+import { GraphQLError } from 'graphql/error';
 import dbConnect from '@shared/db/db';
-import { Label } from '@shared/db/model';
+import requireUser from '@shared/db/graphql/utils/requireUser';
+import { Board, Label } from '@shared/db/model';
+import { MutationResolvers, QueryResolvers } from '@shared/types/generated/graphql';
 
 export const labelResolvers = {
     Query: {
-        getLabels: async (_: any, { board }: { board: string }) => {
+        getLabels: (async (_, { boardId }, ctx) => {
+            const userId = requireUser(ctx?.user);
             await dbConnect();
-            return Label.find({ board });
-        },
+
+            return Label.find({ board: boardId, userId });
+        }) satisfies QueryResolvers['getLabels'],
     },
     Mutation: {
-        createLabel: async (
-            _: any,
-            {
-                name,
-                color,
-                order,
-                board,
-            }: { name: string; color: string; order: number; board: string },
-        ) => {
+        createLabel: (async (_, { name, color, order, boardId }, ctx) => {
+            const userId = requireUser(ctx?.user);
             await dbConnect();
-            return Label.create({ name, color, order, board });
-        },
-        deleteLabel: async (_: any, { id }: { id: string }) => {
-            await dbConnect();
-            await Label.deleteOne({ _id: id });
 
+            const board = await Board.findOne({ _id: boardId, userId });
+            if (!board) {
+                throw new GraphQLError('This board does not exist', {
+                    extensions: { code: 'FORBIDDEN' },
+                });
+            }
+
+            return Label.create({ name, color, order, board: boardId, userId });
+        }) satisfies MutationResolvers['createLabel'],
+        deleteLabel: (async (_, { id }, ctx) => {
+            const userId = requireUser(ctx?.user);
+            await dbConnect();
+
+            await Label.deleteOne({ _id: id, userId });
             return id;
-        },
-        updateLabel: async (
-            _: any,
-            { id, name, color }: { id: string; name: string; color: string },
-        ) => {
+        }) satisfies MutationResolvers['deleteLabel'],
+        updateLabel: (async (_, { id, name, color }, ctx) => {
+            const userId = requireUser(ctx?.user);
             await dbConnect();
-            await Label.updateOne({ _id: id }, { $set: { name, color } });
 
-            return Label.findById(id);
-        },
-        updateLabelsOrders: async (
-            _: any,
-            { labels }: { labels: { id: string; order: number }[] },
-        ) => {
+            await Label.updateOne({ _id: id, userId }, { $set: { name, color } });
+            return Label.findOne({ _id: id, userId });
+        }) satisfies MutationResolvers['updateLabel'],
+        updateLabelsOrders: (async (_, { labels, boardId }, ctx) => {
+            const userId = requireUser(ctx?.user);
             await dbConnect();
 
             await Promise.all(
-                labels.map(l => Label.updateOne({ _id: l.id }, { $set: { order: l.order } })),
+                labels.map(l =>
+                    Label.updateOne({ _id: l.id, userId }, { $set: { order: l.order } }),
+                ),
             );
 
-            return Label.find();
-        },
+            return Label.find({ boardId, userId });
+        }) satisfies MutationResolvers['updateLabelsOrders'],
     },
 };
